Bumping version to 3.2.14.rc.005. Previous Commit message: "New filter to customize retrieved event list (#10)
* New filter to customize retrieved event list

New filter to customize the retrieved event list, allowing full customization of the list obtained from the database, allowing the dynamic generation of events, or the removal of specific events.

Also, allowing the cancellation of the insertion of a event into the JSON if the already existing filter FHEE__EE_Calendar__get_calendar_events__calendar_datetime returns false.

* Update EED_Espresso_Calendar.module.php

Co-authored-by: Brent Christensen <brent@eventespresso.com>

* Removing space at the end of line

Co-authored-by: Brent Christensen <brent@eventespresso.com>"